% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc1_draws.R
\name{auroc1}
\alias{auroc1}
\title{Calculate area under the empirical pseudo-type 1 receiver operating characteristic curve}
\usage{
auroc1(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL
)
}
\arguments{
\item{data}{The data frame to aggregate}

\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{auroc1}: the area under the pseudo type 1 ROC curve
}
}
\description{
Calculate area under the empirical pseudo-type 1 receiver operating characteristic curve
}
\examples{
# calculate area under the type 1 ROC
auroc1(example_data())

# calculate type 1 ROCs by condition
auroc1(sim_metad_condition(), condition)
}
\seealso{
\code{\link[=roc1]{roc1()}}, \code{\link[=auroc1_draws]{auroc1_draws()}}, \code{\link[=auroc1_rvars]{auroc1_rvars()}}
}
