% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc1_draws.R
\name{auroc1_draws}
\alias{auroc1_draws}
\alias{add_auroc1_draws}
\alias{auroc1_rvars}
\alias{add_auroc1_rvars}
\title{Obtain posterior draws of the area under the pseudo type 1 receiver operating
characteristic (ROC) curve.}
\usage{
auroc1_draws(
  object,
  newdata,
  ...,
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response"
)

add_auroc1_draws(newdata, object, ...)

auroc1_rvars(
  object,
  newdata,
  ...,
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response"
)

add_auroc1_rvars(newdata, object, ...)
}
\arguments{
\item{object}{The \code{brms} model with the \code{metad} family}

\item{newdata}{A data frame from which to generate posterior predictions}

\item{...}{Additional parameters passed to \link[tidybayes:epred_draws]{tidybayes::epred_draws} or
\link[tidybayes:epred_rvars]{tidybayes::epred_rvars}}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}
}
\value{
a tibble containing posterior draws of the area under the pseudo
type 1 ROC with the following columns:
\itemize{
\item \code{.row}: the row of \code{newdata}
\item \code{.chain}, \code{.iteration}, \code{.draw}: for \code{auroc1_draws} and \code{add_auroc1_draws}, identifiers for the posterior sample
\item \code{auroc1}: the area under the pseudo type 1 ROC curve
}
}
\description{
Given a data frame and a meta-d' model, adds estimates of the
area under the type 1 ROC curve. For \code{auroc1_draws} and \code{add_auroc1_draws},
estimates are returned in a tidy tibble with one row per posterior draw.
For \code{auroc1_rvars} and \code{add_auroc1_rvars}, parameters are returned as
\link[posterior:rvar]{posterior::rvar}s, with one row per row in \code{newdata}.
}
\examples{
\donttest{
newdata <- tidyr::tibble(.row = 1)

# compute pseudo-type 1 ROC curve
# equivalent to `auroc1_draws(example_model(), newdata)`
add_auroc1_draws(newdata, example_model())

# use posterior::rvar for additional efficiency
# equivalent to `add_auroc1_draws(newdata, example_model())`
auroc1_rvars(example_model(), newdata)
}
}
\seealso{
\code{\link[=auroc1]{auroc1()}}, \code{\link[=roc1_draws]{roc1_draws()}}, \code{\link[=roc1_rvars]{roc1_rvars()}}
}
