% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc2_draws.R
\name{auroc2}
\alias{auroc2}
\title{Calculate the area under empirical type 2 receiver operating characteristic curves}
\usage{
auroc2(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL,
  by_response = TRUE
)
}
\arguments{
\item{data}{The data frame to aggregate}

\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}

\item{by_response}{If \code{TRUE} (default), calculate type 2 ROCs conditional on
type 1 response.}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{{.response}} (if \code{by_response=TRUE}): the type 1 response
\item \code{auroc2}: the area under the type 2 ROC
}
}
\description{
Calculate the area under empirical type 2 receiver operating characteristic curves
}
\examples{
# calculate type 2 ROCs by stimulus
auroc2(example_data())

# calculate type 2 ROCs by condition, averaging over type 1 responses
auroc2(sim_metad_condition(), condition, by_response = FALSE)
}
\seealso{
\code{\link[=roc2]{roc2()}}, \code{\link[=auroc2_draws]{auroc2_draws()}}, \code{\link[=auroc2_rvars]{auroc2_rvars()}}
}
