% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cov_matrix}
\alias{cov_matrix}
\title{Generate a covariance matrix.}
\usage{
cov_matrix(S, OMEGA)
}
\arguments{
\item{S}{A vector of standard deviations}

\item{OMEGA}{A correlation matrix}
}
\value{
an \eqn{N \times N} covariance matrix, where \code{N = length(S)}.
}
\description{
Generate a covariance matrix.
}
\examples{
sds <- c(1, 2)
corrs <- matrix(c(1, .5, .5, 1), nrow = 2)
cov_matrix(sds, corrs)
}
