% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_data}
\alias{example_data}
\title{Simulated data for example model fitting}
\usage{
example_data()
}
\value{
A tibble of 1000 observations containing the following columns:
\itemize{
\item \code{trial}: the trial number
\item \code{stimulus}: the stimulus presence (\code{0} or \code{1})
\item \code{response}: the simulated type 1 response
\item \code{confidence}: the simulated type 2 response
\item \code{correct}: the accuracy of the simulated type 1 response
\item \code{dprime}, \code{c}, \code{meta_dprime}, \code{M}, \code{meta_c2_0}, \code{meta_c2_1}: the
parameters of the model used for simulation
\item \code{theta}, \code{theta_1}, \code{theta_2}: the joint, type 1, and type 2 response
probabilities of the model used for simulation
}
}
\description{
A simulated data set of 1000 trials from a two-alternative
forced choice task with 4 levels of confidence.
}
\examples{
# Fit an empty model on the example data
# (remove empty=TRUE to actually fit the model)
fit_metad(N ~ 1, example_data(), empty = TRUE)
}
\seealso{
\code{\link[=sim_metad]{sim_metad()}}
}
