% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_model}
\alias{example_model}
\title{Example meta-d' model for model post-processing}
\usage{
example_model()
}
\value{
A \code{brmsfit} object fitted to simulated data
}
\description{
A model fit to the simulated data \link{example_data}. This
model includes one constant set of parameters, with no multilevel
structure.
}
\examples{
\donttest{
# inspect summary of posterior distribution
summary(example_model())

# obtain posterior expectations
epred_draws_metad(example_model(), tidyr::tibble(.row = 1))
}
}
\seealso{
\code{\link[=fit_metad]{fit_metad()}}
}
