% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{get_ll}
\alias{get_ll}
\title{Get the likelihood parameterization in \code{model}}
\usage{
get_ll(model)
}
\arguments{
\item{model}{The \code{brms} model to get the parameterization for}
}
\value{
A character vector, either \code{"multinomial"} or \code{"categorical"}.
}
\description{
Get the likelihood parameterization in \code{model}
}
\keyword{internal}
