% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{get_stimulus}
\alias{get_stimulus}
\title{Get the stimulus variable in \code{model}}
\usage{
get_stimulus(model)
}
\arguments{
\item{model}{The \code{brms} model to get the variable for}
}
\value{
A character vector.
}
\description{
Get the stimulus variable in \code{model}
}
\keyword{internal}
