% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmetad-package.R
\docType{package}
\name{hmetad-package}
\alias{hmetad}
\alias{hmetad-package}
\title{hmetad: Fit the Meta-D' Model of Confidence Ratings Using 'brms'}
\description{
Implementation of Bayesian regressions over the meta-d' model of psychological data from two alternative forced choice tasks with ordinal confidence ratings. For more information, see Maniscalco & Lau (2012) \doi{10.1016/j.concog.2011.09.021}. The package is a front-end to the 'brms' package, which facilitates a wide range of regression designs, as well as tools for efficiently extracting posterior estimates, plotting, and significance testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metacoglab.github.io/hmetad/}
  \item \url{https://github.com/metacoglab/hmetad}
  \item Report bugs at \url{https://github.com/metacoglab/hmetad/issues}
}

}
\author{
\strong{Maintainer}: Kevin O'Neill \email{kevin.o'neill@ucl.ac.uk} (\href{https://orcid.org/0000-0001-7401-9802}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kevin O'Neill \email{kevin.o'neill@ucl.ac.uk} (\href{https://orcid.org/0000-0001-7401-9802}{ORCID}) [copyright holder]
  \item Stephen Fleming \email{stephen.fleming@ucl.ac.uk} (\href{https://orcid.org/0000-0003-0233-4891}{ORCID}) [copyright holder]
}

}
\keyword{internal}
