% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epred_draws_metad.R
\name{joint_probabilities}
\alias{joint_probabilities}
\title{Calculate empirical joint type 1/type 2 response probabilities}
\usage{
joint_probabilities(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL,
  by_stimulus = TRUE
)
}
\arguments{
\item{data}{The data frame to aggregate}

\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}

\item{by_stimulus}{If \code{TRUE} (default), calculate type 2 response
probabilities conditional on stimulus.}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{{.stimulus}} (if \code{by_stimulus=TRUE}): the stimulus
\item \code{{.response}}: the type 1 response
\item \code{{.confidence}}: the type 2 response
\item \code{{.joint_response}}: the joint type 1/type 2 response
\item \code{n}: the number of rows in \code{data} with the corresponding \code{stimulus} (if \code{by_stimulus=TRUE}), \code{response}, \code{confidence}, and \code{joint_response}
\item \code{p}: the proportion of rows in \code{data} with the corresponding \code{response} (per \code{stimulus} if \code{by_stimulus=TRUE})
}
}
\description{
Given a dataset \code{data}, determine the probability of each combination of type
1 and type 2 responses, optionally conditional on stimulus.
}
\examples{
# calculate type 2 response probabilities by stimulus
joint_probabilities(example_data())

# calculate type 2 response probabilities by condition, averaging over stimuli
joint_probabilities(sim_metad_condition(), condition, by_stimulus = FALSE)
}
\seealso{
\code{\link[tidybayes:epred_draws]{epred_draws()}}, \code{\link[tidybayes:epred_rvars]{epred_rvars()}}
}
