% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_confidence_draws.R
\name{mean_confidence}
\alias{mean_confidence}
\title{Calculate empirical mean confidence}
\usage{
mean_confidence(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL,
  by_stimulus = TRUE,
  by_response = TRUE,
  by_correct = FALSE
)
}
\arguments{
\item{data}{The data frame to aggregate}

\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}

\item{by_stimulus}{If \code{TRUE} (default), calculate mean confidence conditional
on stimulus. Ignored if \code{by_correct=TRUE}.}

\item{by_response}{If \code{TRUE} (default), calculate mean confidence conditional
on type 2 response. Ignored if \code{by_correct=TRUE}.}

\item{by_correct}{If \code{FALSE} (default), calculate mean confidence conditional
on stimulus and/or type 1 response. If \code{TRUE}, instead calculate mean
confidence conditional on accuracy.}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{{.stimulus}}: the stimulus (if \code{by_stimulus=TRUE})
\item \code{{.response}}: the type 1 response (if \code{by_response=TRUE})
\item \code{correct}: the accuracy (if \code{by_correct=TRUE})
\item \code{mean_confidence}: the mean confidence rating
}
}
\description{
Given a dataset \code{data}, determine the mean confidence rating, optionally
conditional on stimulus and/or type 1 response.
}
\examples{
# calculate mean confidence by stimulus and response
mean_confidence(example_data())

# calculate mean confidence by accuracy
mean_confidence(example_data(), by_correct = TRUE)

# calculate mean confidence by condition, averaging over type 1 responses
mean_confidence(sim_metad_condition(), condition, by_response = FALSE)
}
\seealso{
\code{\link[=mean_confidence_draws]{mean_confidence_draws()}}, \code{\link[=mean_confidence_rvars]{mean_confidence_rvars()}}
}
