% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{metac2_parameters}
\alias{metac2_parameters}
\title{Obtain a vector of the names of the \code{K-1} parameters representing the
differences between successive confidence criteria for the meta-d' model with
\code{K} levels of confidence.}
\usage{
metac2_parameters(K, response = "both")
}
\arguments{
\item{K}{The number of confidence levels}

\item{response}{If "both", list confidence criteria parameters for both "0"
and "1" responses. If "zero" or "0", list only confidence criteria for the
"0" response. If "one" or "1", list only confidence criteria for the "1"
response.}
}
\description{
Obtain a vector of the names of the \code{K-1} parameters representing the
differences between successive confidence criteria for the meta-d' model with
\code{K} levels of confidence.
}
\examples{
# list confidence criteria parameters for K=3 confidence levels
metac2_parameters(K = 3)

# list parameters for "0" responses
metac2_parameters(K = 3, response = "zero")

# useful for setting model priors
set_prior("normal(0, 1)", class = "b", dpar = metac2_parameters(K = 4))

}
