% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{normal_lcdf}
\alias{normal_lcdf}
\alias{normal_lccdf}
\title{Normal cumulative distribution functions}
\usage{
normal_lcdf(x, mu)

normal_lccdf(x, mu)
}
\arguments{
\item{x}{The quantile to evaluate the l(c)cdf at}

\item{mu}{The mean of the normal distribution}
}
\value{
\eqn{log(P(X < x))} (for \code{normal_lcdf}) or \eqn{log(P(X > x))} (for \code{normal_lccdf}) where \eqn{X} is sampled from a normal distribution
with mean \code{mu} and standard deviation of \eqn{1}
}
\description{
Normal cumulative distribution functions
}
\examples{
normal_lcdf(0, mu = 1)
normal_lccdf(0, mu = 1)

}
