% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{response_probabilities}
\alias{response_probabilities}
\title{Compute joint response probabilities from aggregated counts}
\usage{
response_probabilities(counts)
}
\arguments{
\item{counts}{A vector (or matrix) of counts of joint type 1/type 2
responses as provided by \link{aggregate_metad}}
}
\value{
A vector (or matrix) of response probabilities \code{P(R, C | S)}
}
\description{
Compute joint response probabilities from aggregated counts
}
\details{
For response \code{R}, confidence \code{C}, stimulus \code{S}, and \code{K=length(counts)/4},
\code{counts} should be a vector (or matrix with rows) of the form:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[N(R=0, C=K, S=0), ..., N(R=0, C=1, S=0),
 N(R=1, C=1, S=0), ..., N(R=1, C=K, S=0),
 N(R=0, C=K, S=1), ..., N(R=0, C=1, S=1),
 N(R=1, C=1, S=1), ..., N(R=1, C=K, S=1)]
}\if{html}{\out{</div>}}

Returns a vector (or matrix with rows) of the form:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[P(R=0, C=K | S=0), ..., P(R=0, C=1 | S=0),
 P(R=1, C=1 | S=0), ..., P(R=1, C=K | S=0),
 P(R=0, C=K | S=1), ..., P(R=0, C=1 | S=1),
 P(R=1, C=1 | S=1), ..., P(R=1, C=K | S=1)]
}\if{html}{\out{</div>}}
}
\examples{
# Aggregate responses from simulated data
d <- sim_metad() |> aggregate_metad()

# Compute conditional response probabilities
response_probabilities(d$N)

# Also works on matrices
matrix(rep(1, 16), nrow = 2) |> response_probabilities()
}
