% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_metad_condition}
\alias{sim_metad_condition}
\title{Simulate from the meta-d' model across separate conditions}
\usage{
sim_metad_condition(
  N_trials = 100,
  dprime = rep(1, 2),
  c = rep(0, 2),
  M = rep(1, 2),
  c2_0_diff = list(rep(0.5, 3), rep(0.5, 3)),
  c2_1_diff = list(rep(0.5, 3), rep(0.5, 3)),
  metac_absolute = TRUE,
  summarize = FALSE,
  lcdf = normal_lcdf,
  lccdf = normal_lccdf
)
}
\arguments{
\item{N_trials}{Total number of trials to simulate. Half of these trials will
have \code{stimulus=0} and half will have \code{stimulus=1}.}

\item{dprime}{The sensitivity of the signal detection agent to simulate}

\item{c}{The response bias of the signal detection agent to simulate}

\item{M}{The metacognitive efficiency of the agent, where negative values
indicate below chance metacognitive sensitivity, \code{0} indicates an absence
of metacognitive sensitivity, values between \code{0} and \code{1} indicate
metacognitive inefficiency, 1\verb{indicates optimal metacognitive sensitivity, and values greater than}1` indicate metacognitive hyper-efficiency.}

\item{c2_0_diff, c2_1_diff}{Distances between confidence thresholds for \code{"0"}
and \code{"1"} responses, such that \code{meta_c2_0 = meta_c - cumsum(c2_0_diff)} and
\code{meta_c2_1 = meta_c + cumsum(c2_1_diff)}.}

\item{metac_absolute}{Determines how to fix the type 1 threshold for modeling
confidence ratings. If \code{metac_absolute=TRUE}, \code{meta_c = c}. Otherwise,
\code{meta_c = M * c}.}

\item{summarize}{Aggregate the data? If \code{summarize=FALSE}, returns a dataset
with one row per observation. If \code{summarize=TRUE}, returns an aggregated
dataset where \code{n} is the number of observations per response, accuracy, and
confidence level.}

\item{lcdf, lccdf}{The log (complement) cumulative distribution function of
the underlying signal distribution. By default, uses a \verb{normal(+/-dprime/2, 1)} distribution.}
}
\value{
A simulated dataset of type 1 responses and confidence ratings, with
columns:
\itemize{
\item \code{trial}: the simulated trial number
\item \code{condition}: the simulated condition number
\item \code{stimulus}: the value of the stimulus on each trial (either \code{0} or \code{1})
\item \code{response}: the simulated type 1 response (either \code{0} or \code{1})
\item \code{correct}: whether \code{stimulus==response} (either \code{0} or \code{1})
\item \code{confidence}: the simulated type 2 response (from \code{1} to \code{length(c2_0_diff)+1})
\item \code{dprime:theta_2}: the simulated agent's parameter values
}

If \code{summarize=TRUE}, the \code{trial} column is replaced with an \code{n} column
indicating the number of simulated type 1/type 2 responses for each
possible value.
}
\description{
Generate a simulated dataset across separate conditions from the
meta-d' model with sensitivity \code{dprime}, response bias \code{c}, metacognitive
efficiency \code{M}, and distances between confidence thresholds \code{c2_0_diff} and
\code{c2_1_diff} (for the two responses).
}
\examples{
sim_metad_condition(N_trials = 10)
sim_metad_condition(N_trials = 10000, summarize = TRUE)
sim_metad_condition(N_trials = 10, c2_0_diff = list(1, .5), c2_1_diff = list(1, .5))
}
