% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type1_draws.R
\name{type1_draws}
\alias{type1_draws}
\alias{add_type1_draws}
\alias{type1_rvars}
\alias{add_type1_rvars}
\title{Calculate posterior draws of type 1 response probabilities}
\usage{
type1_draws(
  object,
  newdata,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  by_stimulus = TRUE
)

add_type1_draws(newdata, object, ...)

type1_rvars(
  object,
  newdata,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  by_stimulus = TRUE
)

add_type1_rvars(newdata, object, ...)
}
\arguments{
\item{object}{The \code{brms} model with the \code{metad} family}

\item{newdata}{A data frame from which to generate posterior predictions}

\item{...}{Additional arguments passed to \link[tidybayes:add_epred_draws]{tidybayes::add_epred_draws} or
\link[tidybayes:add_epred_rvars]{tidybayes::add_epred_rvars}}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{by_stimulus}{If \code{TRUE} (default), calculate conditional type 1 response
probabilities \eqn{P(R=r \vert S=s)}. Otherwise, calculate unconditional
response probabilities \eqn{P(R=r)} as an unweighted average over stimuli.}
}
\value{
a tibble containing posterior draws of model parameters with the
following columns:
\itemize{
\item \code{.row}: the row of \code{newdata}
\item \code{.chain}, \code{.iteration}, \code{.draw}: for \code{epred_draws_metad}, identifiers for the posterior sample
\item \code{{.stimulus}}, \code{{.response}}: identifiers for the response type
\item \code{.epred}: probability of the type 1 response (optionally given the stimulus)
}
}
\description{
Given a data frame and a meta-d' model, adds estimates of type 1
response probabilities (i.e., \eqn{P(R=r \vert S=s)} or \eqn{P(R=r)} for
type 1 response \eqn{R} and stimulus \eqn{S}). For \code{type1_draws_metad} and
\code{add_type1_draws_metad}, estimates are returned in a tidy tibble with one
row per posterior draw. For \code{type1_rvars_metad} and
\code{add_type1_rvars_metad}, parameters are returned as \link[posterior:rvar]{posterior::rvar}s,
with one row per row in \code{newdata}.
}
\examples{
\donttest{
newdata <- tidyr::tibble(.row = 1)

# obtain model predictions
# equivalent to `add_type1_draws(newdata, example_model())`
type1_draws(example_model(), newdata)

# obtain model predictions (`posterior::rvar`)
# equivalent to `add_type1_rvars(newdata, example_model(), by_stimulus = FALSE)`
type1_rvars(example_model(), newdata, by_stimulus = FALSE)
}
}
\seealso{
\code{\link[=type1_probabilities]{type1_probabilities()}}
}
