% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type1_draws.R
\name{type1_probabilities}
\alias{type1_probabilities}
\title{Calculate empirical type 1 response probabilities}
\usage{
type1_probabilities(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL,
  by_stimulus = TRUE
)
}
\arguments{
\item{data}{The data frame to aggregate}

\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}

\item{by_stimulus}{If \code{TRUE} (default), calculate conditional type 1 response
probabilities \eqn{P(R=r \vert S=s)}. Otherwise, calculate unconditional
response probabilities \eqn{P(R=r)} as an unweighted average over stimuli.}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{{.stimulus}} (if \code{by_stimulus=TRUE}): the stimulus
\item \code{{.response}}: the type 1 response
\item \code{n}: the number of rows in \code{data} with the corresponding \code{stimulus} (if \code{by_stimulus=TRUE}) and \code{response}
\item \code{p}: the proportion of rows in \code{data} with the corresponding \code{response} (per \code{stimulus} if \code{by_stimulus=TRUE})
}
}
\description{
Given a dataset \code{data}, determine the probability of each type 1 response,
optionally conditional on stimulus.
}
\examples{
# calculate response probabilities by stimulus
type1_probabilities(example_data())

# calculate response probabilities by condition, averaging over stimuli
type1_probabilities(sim_metad_condition(), condition, by_stimulus = FALSE)
}
\seealso{
\code{\link[=type1_draws]{type1_draws()}}
}
