% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type2_draws.R
\name{type2_draws}
\alias{type2_draws}
\alias{add_type2_draws}
\alias{type2_rvars}
\alias{add_type2_rvars}
\title{Calculate posterior draws of type 2 response probabilities}
\usage{
type2_draws(
  object,
  newdata,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  by_stimulus = TRUE,
  by_response = TRUE,
  by_correct = FALSE
)

add_type2_draws(newdata, object, ...)

type2_rvars(
  object,
  newdata,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  by_stimulus = TRUE,
  by_response = TRUE,
  by_correct = FALSE
)

add_type2_rvars(newdata, object, ...)
}
\arguments{
\item{object}{The \code{brms} model with the \code{metad} family}

\item{newdata}{A data frame from which to generate posterior predictions}

\item{...}{Additional arguments passed to \link[tidybayes:add_epred_draws]{tidybayes::add_epred_draws} or
\link[tidybayes:add_epred_rvars]{tidybayes::add_epred_rvars}}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{by_stimulus}{If \code{TRUE} (default), calculate type 2 response
probabilities separately by stimulus. Otherwise, calculate unconditional
type 2 response probabilities as an unweighted average over stimuli.}

\item{by_response}{If \code{TRUE} (default), calculate type 2 response
probabilities separately by type 1 response. Otherwise, calculate
unconditional type 2 response probabilities as an unweighted average over
type 1 responses.}

\item{by_correct}{If \code{FALSE} (default), calculate type 2 response
probabilities conditional on stimulus and/or type 1 response. If \code{TRUE},
instead calculate probabilities conditional on accuracy.}
}
\value{
a tibble containing posterior draws of model parameters with the
following columns:
\itemize{
\item \code{.row}: the row of \code{newdata}
\item \code{.chain}, \code{.iteration}, \code{.draw}: for \code{epred_draws_metad}, identifiers for the posterior sample
\item \code{{.stimulus}}, \code{{.response}}, \code{{.confidence}}: identifiers for the response type
\item \code{.epred}: probability of the type 1 and type 2 response given the stimulus, \eqn{P(R, C \;\vert\; S)}
}
}
\description{
Given a data frame and a meta-d' model, adds estimates of type 2
response probabilities (i.e., \eqn{P(C=c \vert S=s, R=r)}, \eqn{P(C=c \vert
  S=s)}, \eqn{P(C=c \vert R=r)} or \eqn{P(C=c)} for stimulus \eqn{S}),
type 1 response \eqn{R}, and type 2 response \eqn{C}. For
\code{type2_draws_metad} and \code{add_type2_draws_metad}, estimates are returned in
a tidy tibble with one row per posterior draw. For \code{type2_rvars_metad} and
\code{add_type2_rvars_metad}, parameters are returned as \link[posterior:rvar]{posterior::rvar}s,
with one row per row in \code{newdata}.
}
\examples{
\donttest{
newdata <- tidyr::tibble(.row = 1)

# obtain model predictions
# equivalent to `add_type2_draws(newdata, example_model())`
type2_draws(example_model(), newdata)

# obtain model predictions (`posterior::rvar`)
# equivalent to `add_type2_rvars(newdata, example_model(), by_stimulus = FALSE)`
type2_rvars(example_model(), newdata, by_stimulus = FALSE)
}
}
\seealso{
\code{\link[=type2_probabilities]{type2_probabilities()}}
}
