% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{aggregate_metad}
\alias{aggregate_metad}
\title{Aggregate \code{data} by \code{response}, \code{confidence}, and other columns}
\usage{
aggregate_metad(data, ..., .response = "N", K = NULL)
}
\arguments{
\item{data}{The dataframe to aggregate}

\item{...}{Grouping columns in \code{data}.
These columns will be converted to factors.}

\item{.response}{The name of the resulting column containing trial counts}

\item{K}{The number of confidence levels in \code{data}.
If \code{NULL}, this is estimated from \code{data}.}
}
\value{
A tibble with one row per combination of the variables in \code{...},
and another column named by the value of \code{.response} containing trial counts.
For \code{K} confidence levels, this will be an \verb{N x K*4} matrix, such that the
columns represent:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[N(stimulus=0, response=0, confidence=K), ..., N(stimulus=0, response=0, confidence=1),
 N(stimulus=0, response=1, confidence=1), ..., N(stimulus=0, response=1, confidence=K),
 N(stimulus=1, response=0, confidence=K), ..., N(stimulus=1, response=0, confidence=1),
 N(stimulus=1, response=1, confidence=1), ..., N(stimulus=1, response=1, confidence=K)]
}\if{html}{\out{</div>}}
}
\description{
Counts number of rows in \code{data} with unique combinations values in the
columns \code{response}, \code{confidence}, and any other columns in \code{...}.
}
\examples{
# aggregate a dataset without grouping factors
d <- sim_metad()
aggregate_metad(d)

# aggregate a dataset with grouping factors
d2 <- sim_metad_condition()
aggregate_metad(d2, condition)

# can also aggregate ignoring grouping factors
aggregate_metad(d2)
}
