% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc2_draws.R
\name{auroc2_draws}
\alias{auroc2_draws}
\alias{add_auroc2_draws}
\alias{auroc2_rvars}
\alias{add_auroc2_rvars}
\title{Obtain posterior draws of the area under the type 2 receiver operating
characteristic (ROC) curve.}
\usage{
auroc2_draws(
  object,
  newdata,
  ...,
  .response = "response",
  .confidence = "confidence",
  by_response = TRUE
)

add_auroc2_draws(newdata, object, ...)

auroc2_rvars(
  object,
  newdata,
  ...,
  .response = "response",
  .confidence = "confidence",
  by_response = TRUE
)

add_auroc2_rvars(newdata, object, ...)
}
\arguments{
\item{object}{The \code{brms} model with the \code{metad} family}

\item{newdata}{A data frame from which to generate posterior predictions}

\item{...}{Additional parameters passed to \link[tidybayes:epred_draws]{tidybayes::epred_draws}}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{by_response}{If \code{TRUE} (default), compute separate ROCs for each type 1
response. Otherwise, average ROCs across both type 1 responses.}
}
\value{
a tibble containing posterior draws of the pseudo type 1 ROC with
the following columns:
\itemize{
\item \code{.row}: the row of \code{newdata}
\item \code{.chain}, \code{.iteration}, \code{.draw}: for \code{auroc2_draws} and \code{add_auroc2_draws}, identifiers for the posterior sample
\item \code{{.response}}: the type 1 response for perceived stimulus presence (\eqn{R \in \{0, 1\}})
\item \code{{.confidence}}: the type 2 confidence response (\eqn{C \in [1, K]})
\item \code{p_fa2}: the cumulative probability of an incorrect response (\eqn{P(C\ge c \;\vert\; R\ne S)})
\item \code{p_hit2}: the cumulative probability of a correct response (\eqn{P(C\ge c \;\vert\; R = S)})
}
}
\description{
Given a data frame and a meta-d' model, adds estimates of AUROC2
(optionally for each type 1 response). For \code{auroc2_draws} and
\code{add_auroc2_draws}, estimates are returned in a tidy tibble with one row
per posterior draw. For \code{auroc2_rvars} and \code{add_auroc2_rvars}, parameters
are returned as \code{posterior::rvar}s, with one row per row in \code{newdata}.
}
\examples{
\donttest{
newdata <- tidyr::tibble(.row = 1)

# compute type 2 ROC curve
# equivalent to `add_auroc2_draws(newdata, example_model())`
auroc2_draws(example_model(), newdata)

# use posterior::rvar for additional efficiency
# equivalent to `add_auroc2_rvars(newdata, example_model())`
auroc2_rvars(example_model(), newdata)
}
}
\seealso{
\code{\link[=auroc2]{auroc2()}}, \code{\link[tidybayes:epred_draws]{tidybayes::epred_draws()}}, \code{\link[tidybayes:epred_rvars]{tidybayes::epred_rvars()}}
}
