% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model}
\alias{example_model}
\title{Example meta-d' model for model post-processing}
\format{
A \code{brmsfit} object
}
\source{
Generated using the code \code{fit_metad(N ~ 1, example_data, iter = 500)}
}
\usage{
example_model
}
\description{
A model fit to the simulated data \link{example_data}. This
model includes one constant set of parameters, with no multilevel
structure.
}
\examples{

# inspect summary of posterior distribution
summary(example_model)

# obtain posterior expectations
epred_draws_metad(example_model, tidyr::tibble(.row = 1))

}
\seealso{
\code{\link[=fit_metad]{fit_metad()}}
}
\keyword{datasets}
