% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc1_draws.R
\name{roc1}
\alias{roc1}
\title{Calculate empirical pseudo-type 1 receiver operating characteristic curves}
\usage{
roc1(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL,
  bounds = FALSE
)
}
\arguments{
\item{data}{The data frame to aggregate}

\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}

\item{bounds}{If \code{TRUE}, include the endpoints of the ROC at \eqn{(0, 0)} and
\eqn{(1, 1)}. Otherwise, the endpoints are excluded.}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{{.response}}: the type 1 response
\item \code{{.confidence}}: the type 2 response
\item \code{{.joint_response}}: the joint type 1/type 2 response
\item \code{n_0}: the number of rows in \code{data} with \code{stimulus=0} and the corresponding joint response
\item \code{n_1}: the number of rows in \code{data} with \code{stimulus=1} and the corresponding joint response
\item \code{p_0}: where \code{stimulus=0}, the proportion of rows in \code{data} with joint response equal to \code{.joint_response}
\item \code{p_1}: where \code{stimulus=1}, the proportion of rows in \code{data} with joint response equal to \code{.joint_response}
\item \code{p_fa}: where \code{stimulus=0}, the proportion of rows in \code{data} with joint response greater than \code{.joint_response}
\item \code{p_hit}: where \code{stimulus=1}, the proportion of rows in \code{data} with joint response greater than \code{.joint_response}
}
}
\description{
Given a dataset \code{data}, determine the cumulative probability of each
combination of type 1 and type 2 responses conditional on stimulus.
}
\examples{
# calculate type 1 ROCs
roc1(example_data())

# calculate type 1 ROCs by condition
roc1(sim_metad_condition(), condition)
}
\seealso{
\code{\link[=roc1_draws]{roc1_draws()}}, \code{\link[=roc1_rvars]{roc1_rvars()}}
}
