% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_draws.R
\name{roc1_draws}
\alias{roc1_draws}
\alias{add_roc1_draws}
\title{Obtain posterior draws of the pseudo type 1 receiver operating characteristic (ROC) curve.}
\usage{
roc1_draws(object, newdata, ..., bounds = FALSE)

add_roc1_draws(newdata, object, ...)
}
\arguments{
\item{object}{The \code{brms} model with the \code{metad} family}

\item{newdata}{A data frame from which to generate posterior predictions}

\item{...}{Additional parameters passed to \link[tidybayes:add_predicted_draws]{tidybayes::epred_draws}}

\item{bounds}{If \code{TRUE}, include the endpoints of the ROC at \verb{(0, 0)} and \verb{(1, 1)}.
Otherwise, the endpoints are excluded.}
}
\value{
a tibble containing posterior draws of the pseudo type 1 ROC with the following
columns:
\itemize{
\item \code{.row}: the row of \code{newdata}
\item \code{.chain}, \code{.iteration}, \code{.draw}: identifiers for the posterior sample
\item \code{joint_response}: the combined type 1 / type 2 response (in \code{1:(2*K)} for \code{K} confidence levels)
\item \code{response}: the type 1 response for perceived stimulus presence
\item \code{confidence}: the type 2 confidence response
\item \code{p_fa}: the cumulative probability of a 'present'/'old' response for \code{stimulus==0}
\item \code{p_hit}: the cumulative probability of a 'present'/'old' response for \code{stimulus==1}
}
}
\description{
Obtain posterior draws of the pseudo type 1 receiver operating characteristic (ROC) curve.
}
\examples{
# running few iterations so example runs quickly, use more in practice
m <- fit_metad(N ~ 1, sim_metad(), chains = 1, iter = 500)
newdata <- tidyr::tibble(.row = 1)

# compute pseudo-type 1 ROC curve
roc1_draws(m, newdata)
add_roc1_draws(newdata, m)

# include the ROC bounds
roc1_draws(m, newdata, bounds = TRUE)
}
