% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc2_draws.R
\name{roc2}
\alias{roc2}
\title{Calculate empirical type 2 receiver operating characteristic curve}
\usage{
roc2(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL,
  bounds = FALSE,
  by_response = TRUE
)
}
\arguments{
\item{data}{The data frame to aggregate}

\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}

\item{bounds}{If \code{TRUE}, include the endpoints of the ROC at \eqn{(0, 0)} and
\eqn{(1, 1)}. Otherwise, the endpoints are excluded.}

\item{by_response}{If \code{TRUE} (default), calculate type 2 ROCs conditional on
type 1 response.}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{{.response}} (if \code{by_response=TRUE}): the type 1 response
\item \code{{.confidence}}: the type 2 response
\item \code{n_0}: the number of rows in \code{data} with \code{stimulus=0} and the corresponding \code{joint_response}
\item \code{n_1}: the number of rows in \code{data} with \code{stimulus=1} and the corresponding \code{joint_response}
\item \code{p_0}: for incorrect trials, the proportion of rows in \code{data} with confidence equal to \code{confidence}
\item \code{p_1}: for correct trials the proportion of rows in \code{data} with confidence equal to \code{confidence}
\item \code{p_fa2}: for incorrect trials, the proportion of rows in \code{data} with confidence greater than \code{confidence}
\item \code{p_hit2}: for correct trials, the proportion of rows in \code{data} with confidence greater than \code{confidence}
}
}
\description{
Given a dataset \code{data}, determine the cumulative probability of each type 2
responses conditional on accuracy, optionally conditional on type 1 response.
}
\examples{
# calculate type 2 ROCs by stimulus
roc2(example_data())

# calculate type 2 ROCs by condition, averaging over type 1 responses
roc2(sim_metad_condition(), condition, by_response = FALSE)
}
\seealso{
\code{\link[=roc2_draws]{roc2_draws()}}, \code{\link[=roc2_rvars]{roc2_rvars()}}
}
