% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type2_draws.R
\name{type2_probabilities}
\alias{type2_probabilities}
\title{Calculate empirical type 2 response probabilities}
\usage{
type2_probabilities(
  data,
  ...,
  .stimulus = "stimulus",
  .response = "response",
  .confidence = "confidence",
  .joint_response = "joint_response",
  K = NULL,
  by_stimulus = TRUE,
  by_response = TRUE
)
}
\arguments{
\item{...}{Grouping columns in \code{data}. These columns will be converted to
factors.}

\item{.stimulus}{The name of "stimulus" column}

\item{.response}{The name of "response" column}

\item{.confidence}{The name of "confidence" column}

\item{.joint_response}{The name of "joint_response" column}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is
estimated from \code{data} using the maximum value of either the confidence
column or joint response column.}

\item{by_stimulus}{If \code{TRUE} (default), calculate type 2 response
probabilities conditional on stimulus.}

\item{by_response}{If \code{TRUE} (default), calculate type 2 response
probabilities conditional on type 1 response.}

\item{object}{The \code{brms} model with the \code{metad} family}

\item{newdata}{A data frame from which to generate posterior predictions}
}
\value{
A tibble with columns:
\itemize{
\item \code{...}: the grouping columns in \code{data}
\item \code{{.stimulus}} (if \code{by_stimulus=TRUE}): the stimulus
\item \code{{.response}} (if \code{by_response=TRUE}): the type 1 response
\item \code{{.confidence}}: the type 2 response
\item \code{{.joint_response}} (if \code{by_response=TRUE}): the joint type 1/type 2 response
\item \code{n}: the number of rows in \code{data} with the corresponding \code{stimulus} (if \code{by_stimulus=TRUE}), \code{response} (if \code{by_response=TRUE}), and \code{confidence}
\item \code{p}: the proportion of rows in \code{data} with the corresponding \code{response} (per \code{stimulus} if \code{by_stimulus=TRUE} and per \code{response} if \code{by_response=TRUE})
}
}
\description{
Given a dataset \code{data}, determine the probability of each type 2 response,
optionally conditional on stimulus and/or type 1 response.
}
\examples{
# calculate type 2 response probabilities by stimulus
type2_probabilities(example_data())

# calculate type 2 response probabilities by condition, averaging over stimuli
type2_probabilities(sim_metad_condition(), condition, by_stimulus = FALSE)
}
\seealso{
\code{\link[=type2_draws]{type2_draws()}}
}
